DDF-4336: Separated into container/presentation; added copyright
