+++ ui/packages/catalog-ui-search/src/main/webapp/component/metacard-interactions/metacard-interactions.view.js
@@ -18,7 +18,7 @@ const CustomElements = require('../../js/CustomElements.js')
 const plugin = require('plugins/metacard-interactions')
 
 import React from 'react'
-import InteractionsView from './metacard-interactions'
+import InteractionsView from '../../react-component/container/metacard-interactions'
 
 const MetacardInteractionsView = Marionette.LayoutView.extend({
   template() {
@@ -26,6 +26,7 @@ const MetacardInteractionsView = Marionette.LayoutView.extend({
       model: this.model,
       el: this.$el,
       extensions: this.getExtensions(),
+      categories: this.getCategories(),
     }
     return <InteractionsView {...props} />
   },
@@ -35,8 +36,21 @@ const MetacardInteractionsView = Marionette.LayoutView.extend({
     change: 'render',
   },
   ui: {},
-  getExtensions: function() {},
-  handleShare: function() {},
+  getExtensions() {},
+  handleShare() {},
+
+  /**
+   * Should return a list of items comprising a 'category' of action links that will be rendered in the drop-down.
+   * Ex.
+   * {`category-name`: [{
+   *  parent: `parent-css-class`,
+   *  dataHelp: `Something helpful here`,
+   *  icon: `icon class`, //should be consistent throughout the category
+   *  linkText: `Text to be rendered for the link`,
+   *  actionHandler: () => `Perform on-click`
+   * }]}
+   */
+  getCategories() {},
 })
 
 module.exports = plugin(MetacardInteractionsView)
